backend/requirements_backend.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
crontab==1.0.1
cryptography==43.0.1
DateTime==5.5
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.0.6
django-allauth==65.10.0
django-cors-headers==4.4.0
django-crontab==0.7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
exceptiongroup==1.2.2
executing==2.1.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
inflection==0.5.1
ipython==8.27.0
jedi==0.19.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
matplotlib-inline==0.1.7
mysqlclient==2.2.6
oauthlib==3.3.1
packaging==25.0
parso==0.8.4
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==6.32.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==5.0.0
PyPDF2==3.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
six==1.16.0
social-auth-core==4.7.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
zope.interface==6.4.post2
